Pom update for runtime error.  Fixed searcher logic. file utility class.
I am using some apache file handling libraries.  Since they were not on
the classpath, it was causing a runtime error.  Added the classpath
element to the build.

Instead of checking for a boolean flag if the user wants to continue, if
the user quits, exit.

New file utility class to do things like get the files from the test
directory.