Refactored
Pulled the file validation code out into a utility class so we don't
write it for each search.   I could have made it a default
SearchStrategy method, but it is a utility method, not part of
searching.  I also don't see how/why I'd want the option for subclasses
to override it.

Renamed StringSearchTest SearchTest as all search methods are going to
be tested here via paramteized tests.