Tried to do logging.  Fixed initialization error on fileLineReader.
I was incorrectly ONLY initializng the reader and buffered reader in
FileLineGetter if they were null.  AKA the 1st time.  WRONG.
Now the user initializes the members of fileLineGetter each time for the
file.

Also slightly improved the return conditions of fileLineGetter.  If it
just doesn't get a string, it returns an error condition that we can
check for. 

Also, I put a simple boolean method in the SearchStrategy class for
checking when we want to stop looking in the file.