Massive commit.
Tried to start log4j logging. Still don't have the .properties and .xml
files set up properly.  The logging only sometimes happens and does not
now.

Added a test for benchmarks of the methods.  Also checks to make sure
they at least agree on the result.

I also had a method that tried to index and search the index.  BUT it
will only work for whole word finds.  Needs work.